core: Parse textures over multiple threads
The vast majority of the time spent on getting the textures off the disk
and into GPU memory, is done so inside stbi. As this is mostly CPU work
and some IO, multiple threads can be used to process textures in
parallel.

Locally, on an i3-6100 with 2 cores and 4 threads, loading the Sponza
scene from the second assignment of EDAN35 would take about 2.1 s (not
counting all the processing done by assimp, about 0.4 s). Less than
0.1 s is spent on meshes, with the rest coming from the textures.

After parallelising the stbi work, the loading of the scene (still not
accounting for assimp) is down to 1.2 s, with about 1.0 s spent on stbi
(done in parallel), and about 200 ms for uploading the textures to the
GPU (performed on a single thread).